requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0

fastapi
uvicorn
pydantic
